Cover new conjecture-module code from tests/conjecture/
The strategy-stack note (data.py) and stateful-steps report (engine.py) were only exercised from tests/cover/test_flakiness.py, which check-conjecture-coverage does not run. Move the strategy-stack test and add a stateful-steps engine test under tests/conjecture/ so that suite reaches 100%.
